# swagger.yaml
openapi: 3.0.0
info:
  title: StopBankir AI PRO Actions API v2
  version: "2.0"
paths:
  /search_practice:
    post:
      summary: Поиск судебной практики
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                region:
                  type: string
      responses:
        '200':
          description: Список найденных дел
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    case_number:
                      type: string
                    court:
                      type: string
                    link:
                      type: string
                    summary:
                      type: string
  /calc_deadlines:
    post:
      summary: Калькулятор процессуальных сроков
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_date:
                  type: string
                  format: date
                days:
                  type: integer
      responses:
        '200':
          description: Дата окончания срока
          content:
            application/json:
              schema:
                type: object
                properties:
                  end_date:
                    type: string
                    format: date
  /check_case_status:
    post:
      summary: Проверка статуса дела
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                case_number:
                  type: string
      responses:
        '200':
          description: Статус дела
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  last_event:
                    type: string
  /check_counterparty:
    post:
      summary: Проверка контрагента по ИНН
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                inn:
                  type: string
      responses:
        '200':
          description: Информация о контрагенте
          content:
            application/json:
              schema:
                type: object
                properties:
                  bankruptcy_status:
                    type: string
                  fedresurs_publications:
                    type: string
                  court_cases_summary:
                    type: string
                  risk_assessment:
                    type: string
  /summarize_pdf:
    post:
      summary: Резюме PDF документа
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Резюме PDF
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
  /compare_positions:
    post:
      summary: Сравнение правовых позиций
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text1:
                  type: string
                text2:
                  type: string
      responses:
        '200':
          description: Результат сравнения
          content:
            application/json:
              schema:
                type: object
                properties:
                  comparison_summary:
                    type: string
  /suggest_documents:
    post:
      summary: Рекомендации по документам
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_role:
                  type: string
                case_stage:
                  type: string
                case_description:
                  type: string
      responses:
        '200':
          description: Список рекомендуемых документов
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggested_documents:
                    type: array
                    items:
                      type: string